Refactor elixir LSP settings
